services/poi_monitor/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
prometheus-client==0.19.0
pyyaml==6.0.1
python-json-logger==2.0.7
